+++ include/fc/filesystem.hpp
@@ -63,7 +63,7 @@ namespace fc {
   void     create_directories( const path& p );
   path     absolute( const path& p );
   path     canonical( const path& p );
-  uint64_t file_size( const path& p );
+  size_t   file_size( const path& p );
   bool     remove( const path& p );
   void     copy( const path& from, const path& to );
