+++ src/error_report.cpp
@@ -83,12 +83,12 @@ fc::string error_frame::to_detail_string()const {
 }
 fc::string error_frame::to_string()const {
     fc::stringstream ss;
-    int64_t prev = 0;
+    size_t prev = 0;
     auto next = desc.find( '$' );
-    while( prev != int64_t(fc::string::npos) && prev < int64_t(desc.size()) ) {
+    while( prev != size_t(fc::string::npos) && prev < size_t(desc.size()) ) {
    //   slog( "prev: %d next %d   %s", prev, next, desc.substr(prev,next).c_str() );
       // print everything from the last pos until the first '$'
-      ss << desc.substr( prev, next-prev );
+      ss << desc.substr( prev, size_t(next-prev) );
 
       // if we got to the end, return it.
       if( next == string::npos ) { return ss.str(); }
@@ -128,7 +128,7 @@ fc::string error_frame::to_string()const {
 }
 fc::string error_report::to_string()const {
   fc::stringstream ss;
-  for( int i = 0; i < stack.size(); ++i ) {
+  for( uint32_t i = 0; i < stack.size(); ++i ) {
     ss << stack[i].to_string() << "\n";
   }
   return ss.str();
